backend/requirements.txt
flask
python-dotenv
gunicorn
supabase
requests
langchain-core
langchain-google-genai
pypdf
langchain-community